Add Stackable Items Overhaul
Overhauls the method I was using to add stackable items (Dirt) to the inventory, among repositioning certain pieces of code for increased modularity. I have not stress-tested this yet, so it may need to be optimised or have bugs fixed, as well as generally cleaning the code for readability.

Does what the title says. This function is activated with a right click + drag. Will require stress testing to account for potential item duplication, but my initial tests indicated that it should be fairly robust.

Basic Functionality:
- Prototype Interface with Stock Assets
- Equip Items
- Move Items around with a single Drag & Drop
- Swap Item positions with Click & Drag
- Items can be made stackable
- Items can be used (no actual functions)
- Splitting Stacks
- Combining Stacks
- Stackable Items are Added to existing stacks, if applicable

Missing Functionality:
- Saving & Loading Functionality
- Bespoke Art for the Interface
- Audio Clips & Sound Triggers
- Controller Support